Databases working.  Single room generation.
There are a few things I want to look into.
First is keeping track of the room structure.  In Unity, I was able to have the room be an empty object with children of floors and walls.  This allowed me to open up new doorways.
If I can't do that, then perhaps my generation routine needs to map out everything on a grid before generating any rooms, and passing in where doorways should go.
I'd much rather be able to have the rooms changable during gameplay.  Breaking a wall and generating what's behind it is fun gameplay.